GoogleAdsense 処理を非クラス化
